requirements.txt
requests
streamlit
pandas
pyyaml

